requirements.txt
requests
bs4
uvicorn
newspaper3k
lxml_html_clean
flask
asgiref
aiohttp